refactor: Use snake case for file names that aren't React components
